+++ banyan/main.py
@@ -16,7 +16,7 @@ from banyan.controllers.role import RoleController
 from banyan.controllers.service import ServiceController
 from banyan.controllers.shield import ShieldController
 from banyan.controllers.user import UserController
-from banyan.controllers.d_resource import DiscoveredResourceController
+from banyan.controllers.discovered_resource import DiscoveredResourceController
 from .core.exc import BanyanError
 
 # configuration defaults
